Added to many features

+++ src/index.ts
@@ -6,12 +6,11 @@ require('dotenv').config();
 const token: string = process.env.TOKEN ?? "<TOKEN>>";
 const guildId: string = process.env.GUILD ?? "<GuildID>>";
 const channelID: string = process.env.CHANNEL ?? "<ChannelID>>";
-const inviteMaxLength: number = 10;
 const repeatSymbolCount: number = 1;
 const repeatEachSymbolCount: number = 4;
 const needInvites: number = 999;
-
-
+const needSaveFancyInvite: boolean = true;
+const totalRepeat: number = 5;
 
 
 
@@ -26,12 +25,15 @@ client.on(`ready`, async () => {
 });
 let attempts: number = 0;
 let founds: number = 0;
+let dontDelete: number = 0;
+let betters: number = 0;
 const keys: string[] = [];
 function startCycle(){
     const interval = setInterval(async () => {
         const invite = await createInvite();
-        const code = invite.code.split('gg/')[1]?.toLowerCase() ?? invite.code.toLowerCase();
-        const response = checkCode(code);
+        const code = invite.code.split('gg/')[1] ?? invite.code;
+        const lower = code.toLowerCase();
+        const response = checkCode(lower);
         attempts++;
         if(response.success){
             clearInterval(interval);
@@ -45,8 +47,15 @@ function startCycle(){
             }else
                 startCycle();
         }else{
-            console.log(`${attempts}.) ${response.diagnostic.count} ${response.diagnostic.repeat} ${invite} ${founds}`);
-            invite.delete();
+            console.log(`${attempts}.) ${response.diagnostic.count} ${response.diagnostic.repeat} ${code} ${founds} ${betters} ${dontDelete}`);
+            if(response.diagnostic.dontDelete && needSaveFancyInvite){
+                dontDelete++;
+            }else if(response.diagnostic.better){
+                betters++;
+            }else{
+                invite.delete();
+            }
+
         }
     }, 4000);
 }
@@ -57,7 +66,9 @@ function checkCode(code: string): ICheckingResponse{
         count: 0,
         diagnostic: {
             repeat: 0,
-            count: 0
+            count: 0,
+            dontDelete: false,
+            better: false
         }
     };
     const symbols: any = {};
@@ -69,7 +80,7 @@ function checkCode(code: string): ICheckingResponse{
     }
     for(let k in symbols){
         if(symbols[k] > 1){
-            response.diagnostic.repeat += symbols[k] -1;
+            response.diagnostic.repeat += symbols[k];
             response.diagnostic.count++;
         }
         if(symbols[k] >= repeatEachSymbolCount)
@@ -78,7 +89,33 @@ function checkCode(code: string): ICheckingResponse{
             response.count++;
         }
     }
-    if(code.length <= inviteMaxLength && response.count >= repeatSymbolCount)
+    if(response.diagnostic.repeat >= 4 && response.diagnostic.count >= 2){
+        const strings: string[] = [];
+        for(let k in symbols){
+            if(symbols[k] > 1){
+                let str: string = "";
+                for(let i = 0; i < symbols[k]; i++){
+                    str += k;
+                }
+                strings.push(str);
+            }
+        }
+
+        if(response.diagnostic.count == 2){
+            if(code.includes(`${strings[0]}${strings[1]}`) || code.includes(`${strings[1]}${strings[0]}`))
+                response.diagnostic.dontDelete = true;
+        }else if(response.diagnostic.count == 3) {
+            if (code.includes(`${strings[0]}${strings[1]}${strings[2]}`) ||
+                code.includes(`${strings[2]}${strings[1]}${strings[0]}`) ||
+                code.includes(`${strings[1]}${strings[0]}${strings[2]}`) ||
+                code.includes(`${strings[2]}${strings[0]}${strings[1]}`))
+                response.diagnostic.dontDelete = true;
+        }
+
+    }
+    if(totalRepeat > 0 && response.diagnostic.repeat >= totalRepeat)
+        response.diagnostic.better = true;
+    if(response.count >= repeatSymbolCount)
         response.success = true;
     return response;
 }
@@ -99,4 +136,9 @@ interface ICheckingResponse{
 interface IDiagnosticInformation{
     repeat: number;
     count: number;
+    dontDelete: boolean;
+    better: boolean;
 }
+
+
+
